enable adding claims to organizations
