+++ geoinfo/models.py
@@ -3,12 +3,9 @@ import json
 from django.db import models
 from django.utils.translation import ugettext as _
 
-from claim.models import Organization, Claim
+from claim.models import Organization
 
 
-# TODO(autogestion) Maybe we have to create separate app,
-# wich would handle geo info and will contain
-# Layer and Polygon models
 class Layer(models.Model):
     """
         Some general representation of map.
@@ -52,47 +49,19 @@ class Polygon(models.Model):
 
     def generate_map_polygon(self):
         orgs = []
-        # claims_count = 0
         for org in self.organizations.all():
-            # orgs.append(org.get_map_representation())
-            # claims_count += org.total_claims
             orgs.append({'id': org.id,
                         'name': org.name,
-                         'claims_count': org.total_claims})
-
-        # names = " <br> ".join([org['name'] for org in orgs])
+                         'claims_count': org.claim_set.all().count()})
 
         return {
             "type": "Feature",
             "properties": {
                 "ID": self.polygon_id,
-                # "NAME": names,
                 "ORGANIZATIONS": orgs,
-                # "CLAIM_COUNT": claims_count
             },
             "geometry": json.loads(self.coordinates)
         }
-
-    # def get_json_claims(self):
-    #     # Q(polygon_id=self.id) jsut compatible layer.
-    #     # Remove me after release. Not critical.
-    #     claims = Claim.objects.filter(
-    #         organization__in=self.organizations.all())
-    #     claims_list = []
-
-    #     if claims:
-    #         for claim in claims:
-    #             username = claim.complainer.username if\
-    #                 claim.complainer else _("Anon")
-    #             claims_list.append({
-    #                 'organization_id': claim.organization.id,
-    #                 'organization_name': claim.organization.name,
-    #                 'text': claim.text,
-    #                 'servant': claim.servant,
-    #                 'complainer': username
-    #             })
-
-    #     return json.dumps(claims_list)
 
     def organization_count(self):
         return self.organizations.all().count()
