+++ locale/ru/LC_MESSAGES/django.po
@@ -8,7 +8,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: PACKAGE VERSION\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2015-09-25 06:49+0000\n"
+"POT-Creation-Date: 2015-09-27 17:37+0000\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language-Team: LANGUAGE <LL@li.org>\n"
@@ -19,160 +19,156 @@ msgstr ""
 "Plural-Forms: nplurals=3; plural=(n%10==1 && n%100!=11 ? 0 : n%10>=2 && n"
 "%10<=4 && (n%100<10 || n%100>=20) ? 1 : 2);\n"
 
-#: claim/management/commands/initiate_db.py:70
-msgid "No name"
-msgstr ""
-
-#: claim/models.py:15
+#: claim/models.py:13
 msgid "Unknown"
 msgstr ""
 
-#: claim/models.py:16
+#: claim/models.py:14
 msgid "Міністерство фінансів України"
 msgstr ""
 
-#: claim/models.py:17
+#: claim/models.py:15
 msgid "Міністерство соціальної політики України"
 msgstr ""
 
-#: claim/models.py:18
+#: claim/models.py:16
 msgid ""
 "Міністерство регіонального розвитку, будівництва та житлово-комунального "
 "господарства України"
 msgstr ""
 
-#: claim/models.py:20
+#: claim/models.py:18
 msgid "Міністерство охорони здоров'я України "
 msgstr ""
 
-#: claim/models.py:21
+#: claim/models.py:19
 msgid "Міністерство освіти і науки України"
 msgstr ""
 
-#: claim/models.py:22
+#: claim/models.py:20
 msgid "Міністерство оборони України"
 msgstr ""
 
-#: claim/models.py:23
+#: claim/models.py:21
 msgid "Міністерство молоді та спорту України"
 msgstr ""
 
-#: claim/models.py:24
+#: claim/models.py:22
 msgid "Міністерство культури України"
 msgstr ""
 
-#: claim/models.py:25
+#: claim/models.py:23
 msgid "Міністерство інфраструктури України"
 msgstr ""
 
-#: claim/models.py:26
+#: claim/models.py:24
 msgid "Міністерство інформаційної політики України"
 msgstr ""
 
-#: claim/models.py:27
+#: claim/models.py:25
 msgid "Міністерство закордонних справ України"
 msgstr ""
 
-#: claim/models.py:28
+#: claim/models.py:26
 msgid "Міністерство енергетики та вугільної промисловості України"
 msgstr ""
 
-#: claim/models.py:30
+#: claim/models.py:28
 msgid "Міністерство економічного розвитку і торгівлі України"
 msgstr ""
 
-#: claim/models.py:31
+#: claim/models.py:29
 msgid "Міністерство екології та природних ресурсів України"
 msgstr ""
 
-#: claim/models.py:32
+#: claim/models.py:30
 msgid "Міністерство внутрішніх справ України"
 msgstr ""
 
-#: claim/models.py:33
+#: claim/models.py:31
 msgid "Міністерство аграрної політики та продовольства України"
 msgstr ""
 
-#: claim/models.py:34
+#: claim/models.py:32
 msgid "Міністерство юстиції України"
 msgstr ""
 
-#: claim/models.py:123 templates/add_page.html:33
-msgid "Organization"
-msgstr ""
-
-#: claim/models.py:124
-msgid "District"
-msgstr ""
-
-#: claim/models.py:125
-msgid "Country"
-msgstr ""
-
-#: claim/models.py:175
+#: claim/models.py:63
 msgid "Anon"
 msgstr ""
 
-#: corruption_tracker/settings.py:90
+#: corruption_tracker/settings.py:91
 msgid "Ukrainian"
 msgstr ""
 
-#: corruption_tracker/settings.py:91
+#: corruption_tracker/settings.py:92
 msgid "English"
 msgstr ""
 
-#: corruption_tracker/settings.py:92
+#: corruption_tracker/settings.py:93
 msgid "Russian"
 msgstr ""
 
-#: corruption_tracker/settings.py:136
+#: corruption_tracker/settings.py:138
 msgid "Streets"
 msgstr ""
 
-#: corruption_tracker/settings.py:139
+#: corruption_tracker/settings.py:141
 msgid "Satellite"
 msgstr ""
 
-#: templates/add_page.html:18
+#: geoinfo/models.py:21 templates/add_page.html:30
+msgid "Organization"
+msgstr ""
+
+#: geoinfo/models.py:22
+msgid "District"
+msgstr ""
+
+#: geoinfo/models.py:23
+msgid "Country"
+msgstr ""
+
+#: templates/add_page.html:15
 msgid ""
 "To add claim please click on building on the map, or enter organization name "
 "in searchbox 'Organization', name of servant and describe what was wrong. "
 "After it press 'Send claim' button."
 msgstr ""
 
-#: templates/add_page.html:26 templates/base.html:35
+#: templates/add_page.html:23 templates/base.html:35
 msgid "Add claim"
 msgstr ""
 
-#: templates/add_page.html:34 templates/home.html:29
+#: templates/add_page.html:31 templates/home.html:26
 msgid "Organization name"
 msgstr ""
 
-#: templates/add_page.html:37
-msgid "Polygon ID"
+#: templates/add_page.html:34
+msgid "Organization ID"
 msgstr ""
 
-#: templates/add_page.html:41
+#: templates/add_page.html:39
 msgid "Servant name"
 msgstr ""
 
-#: templates/add_page.html:45
+#: templates/add_page.html:43
 msgid "Claim message"
 msgstr ""
 
-#: templates/add_page.html:54
+#: templates/add_page.html:52
 msgid "Keep me anonymously"
 msgstr ""
 
-#: templates/add_page.html:68
+#: templates/add_page.html:66
 msgid "Send claim"
 msgstr ""
 
-#: templates/add_page.html:130
+#: templates/add_page.html:127
 msgid "Too much claims. Please wait an hour and try again."
 msgstr ""
 
-#: templates/add_page.html:134
+#: templates/add_page.html:131
 msgid "Thank you for your message."
 msgstr ""
 
@@ -215,32 +211,32 @@ msgstr ""
 msgid "Register"
 msgstr ""
 
-#: templates/home.html:18
+#: templates/home.html:15
 msgid ""
 "To check claims you should click on building on the map, or enter "
 "organization name in searchbox."
 msgstr ""
 
-#: templates/home.html:26 templates/home.html.py:31
+#: templates/home.html:23 templates/home.html.py:28
 msgid "Search"
 msgstr ""
 
-#: templates/home.html:43 templates/home.html.py:93
+#: templates/home.html:40 templates/home.html.py:90
 msgid "Complainer"
 msgstr ""
 
-#: templates/home.html:44 templates/home.html.py:92
+#: templates/home.html:41 templates/home.html.py:89
 msgid "Servant"
 msgstr ""
 
-#: templates/home.html:45 templates/home.html.py:94
+#: templates/home.html:42 templates/home.html.py:91
 msgid "Claim text"
 msgstr ""
 
-#: templates/home.html:60
+#: templates/home.html:57
 msgid "Name not found"
 msgstr ""
 
-#: templates/home.html:108
+#: templates/home.html:105
 msgid "No claims for this polygon"
 msgstr ""
