+++ static/js/map_init.js
@@ -1,23 +1,17 @@
 function main_map_init (map, options) {
-    // Add GeoJSON layer
-    var marker;
+    // Add GeoJSON layer   
+    var marker, org_id;
 
     // Add building markers with popups to buildings.
-    for (var i = buildings['features'].length - 1; i >= 0; i--) {
-        // polygon_id = buildings['features'][i]['properties']['ID'];
-
-        var org_id;
+    for (var i = buildings['features'].length - 1; i >= 0; i--) {        
 
         for (var ii = buildings['features'][i]['properties']["ORGANIZATIONS"].length - 1; ii >= 0; ii--) {
-
             org_id = buildings['features'][i]['properties']["ORGANIZATIONS"][ii]['id'];
-
 
             var myIcon = L.divIcon({
                 className: 'icon_with_number',
                 html: buildings['features'][i]['properties']["ORGANIZATIONS"][ii]['claims_count']
             });
-
 
             marker = L.marker(
                 [
@@ -39,14 +33,5 @@ function main_map_init (map, options) {
     };
 
     L.geoJson(buildings).addTo(map);
-
-    // // create location control and add to map
-    // lc = L.control.locate({
-    //         follow: true,
-    //         strings: {
-    //             title: "подпись"
-    //         }
-    //     }).addTo(map);
-    //  // request location update and set location
-    // lc.locate();                       
+                    
 }
