+++ templates/home.html
@@ -5,10 +5,7 @@
 {% block head %}
     <title>Corruption map</title>
     {% leaflet_js %}  
-    {% leaflet_css %}<!-- 
-    <script src="http:///rawgithub.com/domoritz/leaflet-locatecontrol/gh-pages/dist/L.Control.Locate.min.js"</script>
-    <link href="http://maxcdn.bootstrapcdn.com/font-awesome/4.2.0/css/font-awesome.min.css" rel="stylesheet">
-	<link href="http:///rawgithub.com/domoritz/leaflet-locatecontrol/gh-pages/dist/L.Control.Locate.min.css" rel="stylesheet"> -->
+    {% leaflet_css %}
 {% endblock %}
